# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maz Delivery API",
        "description": "Maz Delivery project",
        "contact": {
            "email": "info@mazdelivery.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.mazdelivery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/request-van": {
            "get": {
                "tags": [
                    "requestVan"
                ],
                "summary": "Get All RequestVan",
                "description": "Get All RequestVan",
                "operationId": "requestVanLists",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "requestVan"
                ],
                "summary": "Post New RequestVan",
                "description": "Post New RequestVan",
                "operationId": "requestVanCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestVanCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/request-van/{id}": {
            "get": {
                "tags": [
                    "requestVan"
                ],
                "summary": "Show an RequestVan",
                "description": "Show an RequestVan",
                "operationId": "requestVanDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "RequestVan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "requestVan"
                ],
                "summary": "Update an RequestVan",
                "description": "Update an RequestVan",
                "operationId": "requestVanUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "RequestVan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestVanUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "requestVan"
                ],
                "summary": "Delete an RequestVan",
                "description": "Delete an RequestVan",
                "operationId": "requestVanDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "RequestVan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RequestVanCreateFormRequest": {
                "title": "Request Van Create Form Request Fields",
                "description": "Request Van Create request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
            "RequestVanUpdateFormRequest": {
                "title": "Van Request Update Form Request Fields",
                "description": "Van Request Update request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}